// hr-frontend/src/lib/api/generated/model/organizationUnitImportPreviewRow.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface OrganizationUnitImportPreviewRow {
  rowNumber?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  unitType?: string;
  unitTypeId?: number;
  parentUnitCode?: string;
  parentOrgUnitId?: number;
  startDate?: string;
  endDate?: string;
  active?: boolean;
  valid?: boolean;
  errors?: string[];
  warnings?: string[];
}