// hr-frontend/src/lib/api/generated/model/organizationUnitImportResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrganizationUnitImportResultRow } from './organizationUnitImportResultRow';

export interface OrganizationUnitImportResult {
  totalRows?: number;
  importedRows?: number;
  skippedRows?: number;
  failedRows?: number;
  rows?: OrganizationUnitImportResultRow[];
}